+++ src/main/java/com/example/CreditApplicationSystem/exception/ClientNotFoundException.java
@@ -2,7 +2,7 @@ package com.example.CreditApplicationSystem.exception;
 
 public class ClientNotFoundException extends RuntimeException {
 
-    public ClientNotFoundException(Integer id) {
+    public ClientNotFoundException(Long id) {
         super("Could not find employee " + id);
     }
 }
